scheduler: prevent nil pointer ref when reschedule policy is missing
When upgrading from older versions of Nomad, the reschedule policy block may be
nil. There is logic to handle this safely in the `NextRescheduleTimeByTime` used
for allocs on disconnected clients, but it's missing from the
`NextRescheduleTime` method used by more typical allocations. Return an empty
time object in this case.

Fixes: https://github.com/hashicorp/nomad/issues/24846