deps: remove gofakeit (#25073)
This dependency is only used to generate mock `Variables`. The only time the
faked values are meaningful would be in the state store and RPC handler tests,
where we are always setting the values directly so that we can control
unblocking behaviors. Remove most of the random generation and remove the
dependency.

Closes: https://github.com/hashicorp/nomad/pull/25066