upgrade tests: add transparent proxy workload
Add an upgrade test workload for Consul service mesh with transparent proxy. In
order to run this workload, we need a service intention, so I've added that to
the `provision-infra` module. This is an example of a workload where simply
being up and running isn't enough to validate successful upgrades, so this
includes a health check that will kill the allocation. We can use that as a
model for other test workloads.

Ref: https://hashicorp.atlassian.net/browse/NET-12217