dynamic host volumes: mark volumes on down nodes as unavailable (#25188)
We update the status of a volume when the node fingerprint changes. But if a
node goes down, we still show the volume as available. The scheduler behavior is
correct because a down node can never have work scheduled on it, but it might be
confusing for job authors who are looking at volumes that are showing as
available.

Update the volume update logic we have on node updates to include marking the
volume as unavailable when a node goes down.

Fixes: https://hashicorp.atlassian.net/browse/NET-12068