generic paginator (#25252)
The paginator was developed before generics were available, so we've had to work
around a lack of compile-time safety by creating configuration objects at
runtime that require a lot of branching and type casts. This results in a lot of
added boilerplate in the RPC handlers.

Refactor the paginator to take advantage of generics.
* Move all decision making around tokenization to compile-time by providing
  pre-built generic functions that close over target tokens.
* Remove the `appendFunc` parameter in lieu of a `Stub` function parameter that
  will accept existing `Stub` functions in most cases (with the addition of an
  extra `error` return value).
* Generally remove boilerplate in the RPC handlers as a result, except where a
  given handler wants more complex filtering.

This doesn't reduce the boilerplate we need at the top of many blocking queries
to define the iterator we want based on arguments, which we're typically doing
to decide upon which memdb index we want. That's a query optimization problem
and way beyond the scope of this PR.