upgrade testing: commit missing volume specification
In #25285 we converted the CSI workload for upgrade testing to use a self-hosted
NFS. But the volume spec name got changed to `volume.hcl` in the process, which
is in our `.gitignore` file for the repo. We missed this during testing because
the file existed locally, but it fails in nightly runs.

Ref: https://github.com/hashicorp/nomad/pull/25285
Ref: https://github.com/hashicorp/nomad-e2e/actions/runs/13703979647/job/38324786351