provide allocrunner hooks with taskenv builders instead of built env
Some of our allocrunner hooks require a task environment for interpolating
values based on the node or allocation. But several of the hooks accept an
already-built environment or builder and then keep that in memory. Both of these
retain a copy of all the node attributes and allocation metadata, which balloons
memory usage until the allocation is GC'd.

While we'd like to look into ways to avoid keeping the allocrunner around
entirely (see #25372), for now we can significantly reduce memory usage by
creating the task environment lazily and not persisting it in the allocrunner
hooks. Note that the `taskenv.Builder` itself carries around a copy of the node
attributes, so we need to provide the hooks with a factory function to create
the builder and then call its `Build` method.

Fixes: https://github.com/hashicorp/nomad/issues/25269
Ref: https://github.com/hashicorp/nomad/issues/25372