rpc: copy all objects sent with inmem codec
A common source of data races in Nomad tests have been the use of the
InmemCodec. At runtime RPCs are made using a msgpack codec which
copies all objects during encoding and decoding. The InmemCodec does no
encoding or decoding and therefore shares pointers between callers and
RPC handlers. This makes it easy to cause data races in tests that
cannot occur at runtime.

This change uses copystructure to deep copy all objects sent through the
InmemCodec to mimic the behavior of runtime encoding/decoding.

While this creates more garbage and "wastes" CPU, this cost is mostly
incurred during testing, not runtime. This seems like a small price to
pay to make it easier to use Go's data race detector.