tests: fixes a few data races in tests (#25455)
* test: use statedb factory

Swapping fields on Client after it has been created is a race.

* test: lock before checking heartbeat state

Fixes races

* test: fix races by copying fsm objects

A common source of data races in tests is when they insert a fixture
directly into memdb and then later mutate the object. Since objects in
the state store are readonly, any later mutation is a data race.

* test: lock when peeking at eval stats

* test: lock when peeking at serf state

* test: lock when looking at stats

* test: fix default eval broker state test

The test was not applying the config callback. In addition the test
raced against the configuration being applied. Waiting for the keyring
to be initialized resolved the race in my testing, but given the high
concurrency of the various leadership subsystems it's possible it may
still flake.