E2E: test eligibility when getting count for system jobs
While debugging #25383 I discovered that the scaling test for system jobs was
failing in a similar way. I also found log entries suggesting that there are
concurrent operations running to set the node eligibility in the middle of this
test. This is most likely the result of a cleanup function that's happening
concurrently.

This patch tests eligiblity for the counts but can't really fix the underlying
problem, as node eligibility API calls could happen between the time of check
and time of use.

Ref: https://github.com/hashicorp/nomad/issues/25383
Ref: https://hashicorp.atlassian.net/browse/NET-12424