nomad ui: 403 error handling
Currently the Nomad UI will always display an error without any body content. The Nomad CLI does handle
this scenario nicely by outputting the body content coming from the API, see:
https://github.com/hashicorp/nomad/blob/95c914624eaa5bc8f8942bbcbea0b54b207efe99/command/job_run.go#L299

The reason this is for us is because we have a deployment approval handler in between clients and Nomad.
Our policy API returns a 403 when approvals haven't been gathered with a descriptive message. Currently
the Nomad UI unfortunately doesn't return any response content, it just raises a static error.

This change returns the ForbiddenError message if its set and doesn't end with `permission denied`, which
is the body content Nomad returns on ACL errors.